auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "beanie>=1.29.0",
    "fastapi>=0.115.11",
    "motor>=3.7.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.10.6",
    "pymongo>=4.11.2",
    "python-jose>=3.4.0",
    "uvicorn>=0.34.0",
]
